Removed the explicit casting as it was causing problems. Removed the dreaded P() macros.  You must now use a modern compiler for building dgd (sorry Amiga!). Also included a filters file and solution file for the windows build project which makes things look a lot nicer in the project explorer.

+++ src/ed/fileio.c
@@ -213,7 +213,7 @@ bool io_save(editbuf *eb, char *fname, Int first, Int last, int append, io *iobu
 	error((char *) NULL);	/* pass on error */
     }
     eb_range(eb, first, last, put_line, FALSE);
-    if (P_write(ffd, buffer, inbuf) != (int) inbuf) {
+    if (P_write(ffd, buffer, inbuf) != inbuf) {
 	error("error while writing file \"/%s\"", filename);
     }
     ec_pop();
